Agregar validaciones

+++ UsuariosCRUD/Controllers/UsuariosController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using UsuariosCRUD.Models;
+using System.Data.Entity.Infrastructure;
 
 namespace UsuariosCRUD.Controllers
 {
@@ -51,23 +52,38 @@ namespace UsuariosCRUD.Controllers
         }
 
         // POST: Usuarios/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // To protect from overposting attacks, enable the specific properties you want to bind to
+   
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind(Include = "Cedula,Nombre,Apellidos,Profesion")] Usuario usuario)
         {
             if (ModelState.IsValid)
             {
-                db.Usuarios.Add(usuario);
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                bool existe = await db.Usuarios.AnyAsync(u => u.Cedula == usuario.Cedula);
+
+                if (existe)
+                {
+                    ModelState.AddModelError("Cedula", "Ya existe un usuario con esa cédula.");
+                    return View(usuario);
+                }
+
+                try
+                {
+                    db.Usuarios.Add(usuario);
+                    await db.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
+                catch (DbUpdateException)
+                {
+                    ModelState.AddModelError("", "Ocurrió un error al guardar el usuario.");
+                }
             }
 
             return View(usuario);
         }
 
-        // GET: Usuarios/Edit/5
+        // GET: Usuarios/Edit/id
         public async Task<ActionResult> Edit(string id)
         {
             if (id == null)
@@ -82,23 +98,29 @@ namespace UsuariosCRUD.Controllers
             return View(usuario);
         }
 
-        // POST: Usuarios/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: Usuarios/Edit/id
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "Cedula,Nombre,Apellidos,Profesion")] Usuario usuario)
         {
             if (ModelState.IsValid)
             {
-                db.Entry(usuario).State = EntityState.Modified;
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                try
+                {
+                    db.Entry(usuario).State = EntityState.Modified;
+                    await db.SaveChangesAsync();
+                    return RedirectToAction("Index");
+                }
+                catch (DbUpdateException)
+                {
+                    ModelState.AddModelError("", "No se pudo actualizar el usuario.");
+                }
             }
+
             return View(usuario);
         }
 
-        // GET: Usuarios/Delete/5
+        // GET: Usuarios/Delete/id
         public async Task<ActionResult> Delete(string id)
         {
             if (id == null)
@@ -113,15 +135,29 @@ namespace UsuariosCRUD.Controllers
             return View(usuario);
         }
 
-        // POST: Usuarios/Delete/5
+        // POST: Usuarios/Delete/id
+
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(string id)
         {
-            Usuario usuario = await db.Usuarios.FindAsync(id);
-            db.Usuarios.Remove(usuario);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
+            try
+            {
+                Usuario usuario = await db.Usuarios.FindAsync(id);
+
+                if (usuario == null)
+                    return HttpNotFound();
+
+                db.Usuarios.Remove(usuario);
+                await db.SaveChangesAsync();
+
+                return RedirectToAction("Index");
+            }
+            catch (DbUpdateException)
+            {
+                ModelState.AddModelError("", "No se pudo eliminar el usuario.");
+                return View();
+            }
         }
 
         protected override void Dispose(bool disposing)
